+++ src/components/rules/RuleProvider.vue
@@ -2,16 +2,17 @@
   <div class="card w-full gap-2 p-2 text-sm">
     <div class="flex flex-1 items-center gap-2">
       <span>{{ index }}.</span>
-      <span class="text-primary">{{ ruleProvider.name }}</span>
+      <span class="text-main">{{ ruleProvider.name }}</span>
       <span class="badge badge-sm bg-base-200">
         {{ ruleProvider.ruleCount }}
       </span>
     </div>
-    <div class="flex h-4 items-center gap-2 text-xs text-base-content/80">
+    <div class="text-base-content/80 flex h-4 items-center gap-2 text-xs">
       <span>{{ ruleProvider.behavior }}</span>
       <span>{{ ruleProvider.vehicleType }}</span>
       <span>{{ $t('updated') }} {{ fromNow(ruleProvider.updatedAt) }}</span>
       <button
+        v-if="ruleProvider.vehicleType !== 'Inline'"
         :class="twMerge('btn btn-circle btn-xs', isUpdating ? 'animate-spin' : '')"
         @click="updateRuleProviderClickHandler"
       >