+++ src/constant/index.ts
@@ -7,6 +7,7 @@ import {
   SwatchIcon,
 } from '@heroicons/vue/24/outline'
 
+export const TEST_URL = 'https://www.gstatic.com/generate_204'
 export const IPV6_TEST_URL = 'https://ipv6.google.com/generate_204'
 export const NOT_CONNECTED = 0
 export enum LANG {
@@ -40,7 +41,7 @@ export enum CONNECTIONS_TABLE_ACCESSOR_KEY {
   SniffHost = 'sniffHost',
   Destination = 'destination',
   DestinationType = 'destinationType',
-  ProxyNodeAddress = 'proxyNodeAddress',
+  RemoteAddress = 'remoteAddress',
 }
 
 export enum TABLE_WIDTH_MODE {
@@ -96,10 +97,13 @@ export enum CONNECTION_TAB_TYPE {
 }
 
 export enum LOG_LEVEL {
-  Info = 'info',
-  Error = 'error',
-  Warning = 'warning',
+  Trace = 'trace',
   Debug = 'debug',
+  Info = 'info',
+  Warning = 'warning',
+  Error = 'error',
+  Fatal = 'fatal',
+  Panic = 'panic',
   Silent = 'silent',
 }
 
@@ -184,3 +188,41 @@ export const DETAILED_CARD_STYLE = [
     CONNECTIONS_TABLE_ACCESSOR_KEY.Close,
   ],
 ]
+
+export const DEFAULT_THEME = {
+  name: 'custom',
+  id: '',
+  '--border': '1px',
+  '--color-base-100': '#ffffff',
+  '--color-base-200': '#fcfcfc',
+  '--color-base-300': '#f2f2f2',
+  '--color-base-content': '#2d2d33',
+  '--color-primary': '#5a3cd2',
+  '--color-primary-content': '#f3efff',
+  '--color-secondary': '#ea4c5a',
+  '--color-secondary-content': '#fff1f2',
+  '--color-accent': '#49c6c1',
+  '--color-accent-content': '#285e66',
+  '--color-neutral': '#1e1e1f',
+  '--color-neutral-content': '#ececec',
+  '--color-info': '#5b90ff',
+  '--color-info-content': '#273c66',
+  '--color-success': '#44c07a',
+  '--color-success-content': '#1d472f',
+  '--color-warning': '#e5a300',
+  '--color-warning-content': '#705322',
+  '--color-error': '#d13a30',
+  '--color-error-content': '#551d1d',
+  '--depth': '0',
+  '--noise': '0',
+  '--radius-box': '1rem',
+  '--radius-field': '0.5rem',
+  '--radius-selector': '1rem',
+  '--size-field': '0.25rem',
+  '--size-selector': '0.25rem',
+  'color-scheme': 'dark',
+  default: false,
+  prefersdark: false,
+}
+
+export type THEME = typeof DEFAULT_THEME